+++ app/config/app.php
@@ -3,6 +3,9 @@
 return [
 
     'debug' => true,
+
+	//set this to boolean true to output executed database queries
+	'db_profiler' => false
     
 ];