+++ changelog.rst
@@ -9,3 +9,4 @@ Release date: Not released
 
 - Added environment based config overriding
 - Added json response to View handler
+- Added database profiler